[dev] inserting hostname into labels
Quite a few graph titles don't have hostnames in them.  This is fine,
until you start displaying graphs for more than one host
simultaneously.  When you do this, it's really not that obvious which
ones belong to which host (for those graphs which don't explicitly
include the hostname in the title).

This is far from complete, as I've so-far only included corrections for
those graphs which I use on a regular basis (I haven't comprehensively
gone through the entire configuration to fix all of them).