[dev] kludgy layout reshuffle
I changed the layout a little to split the panes along a vertical line
rather than a horizontal one.  I want the listboxes to be able to grow
larger without pushing the graphs down.  For now, I've doubled the fixed
height of the listboxes until I can figure-out how to do it as a
proportion or percentage.

I may swap the panes at some point.